requirements.txt
# Rail Vision 3D — Dependencies
# Core
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0

# 3D Visualization
open3d>=0.17.0

# Scientific Computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
